scrapers/requirements.txt
alembic==1.4.3
attrs==20.2.0
Automat==20.2.0
beautifulsoup4==4.9.0
cffi==1.14.3
constantly==15.1.0
cryptography==3.1.1
cssselect==1.1.0
docutils==0.15.2
fake-useragent==0.1.11
hyperlink==20.0.1
idna==2.10
incremental==17.5.0
jmespath==0.10.0
lxml==4.5.2
Mako==1.1.3
MarkupSafe==1.1.1
pandas
parsel==1.6.0
Protego==0.1.16
psycopg2==2.8.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyDispatcher==2.0.5
PyHamcrest==2.0.2
pyOpenSSL==19.1.0
python-dateutil==2.8.1
python-dotenv
python-editor==1.0.4
queuelib==1.5.0
Scrapy==2.0.1
scrapy-crawlera==1.7.0
scrapy-fake-useragent==1.2.0
service-identity==18.1.0
six==1.15.0
soupsieve==2.0.1
SQLAlchemy==1.3.19
Twisted==20.3.0
Unidecode==1.1.1
urllib3==1.25.10
w3lib==1.22.0
zope.interface==5.1.0
